+++ packages/orq-rc/RELEASES.md
@@ -4459,3 +4459,13 @@ Based on:
 - [typescript v3.14.0-rc.46] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.46] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.46 - packages/orq-rc
+
+## 2025-10-22 13:08:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.638.0 (2.728.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.47] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.47] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.47 - packages/orq-rc